+++ packages/spry_fsrouter/example/app/api/settings/handler.dart
@@ -1,0 +1,5 @@
+import 'package:spry/spry.dart';
+
+handler(Context context) {
+  context.response.send('child(/api/settings)');
+}