+++ packages/spry_fsrouter/example/spry_fsrouter_example.dart
@@ -1,6 +1,0 @@
-import 'package:spry_fsrouter/spry_fsrouter.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}