+++ packages/spry_router/example/main.dart
@@ -21,6 +21,14 @@ void main() async {
       ..send('Hello $name!');
   });
 
+  final Router demo = Router('/demo');
+  demo.all('/', (context) => context.response.send('demo'));
+
+  router.mount('/demo', demo);
+  router.mount('/test', (context) {
+    context.response.send('test');
+  });
+
   await spry.listen(router, port: 3000);
 
   print('Listening on http://localhost:3000');